server/requirements.txt
aiohttp
rpi_ws281x
RPI.GPIO
adafruit-blinka
adafruit-circuitpython-neopixel
holidays
